*refactoring of ConvertShapeNode.
Move ConvertShapeNode from OnnxRuntime.h into ConvertShapeNode.cpp.
