+++ prj/vs2022/CvtOnnx.vcxproj.filters
@@ -261,6 +261,9 @@
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertScatterNdNode.cpp">
       <Filter>OnnxRuntime\S</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertShapeNode.cpp">
+      <Filter>OnnxRuntime\S</Filter>
+    </ClCompile>
     <ClCompile Include="..\..\src\Cvt\OnnxRuntime\ConvertSignNode.cpp">
       <Filter>OnnxRuntime\S</Filter>
     </ClCompile>
